Feature/my plant top bar (#12)
* adding dinamic topbar

* some refactors
